Results of Running the Website Crawl Script
websites_full_test.csv is a sub-result, showing the company websites that did and did not get found in the crawl. 

failed_to_get_website is the failed items in websites_full_test that is crossed with the wikipedia data to get websites and attempt another metadata crawl. 

metadata_on_accepted_sites.csv is the key dataset in the modeling.